process/requirements.txt
h5py==2.9.0
matplotlib==3.1.0
memory-profiler==0.55.0
numpy==1.16.4
opencv-contrib-python-headless==4.1.0.25
pandas==0.23.4
Pillow==6.0.0
psycopg2==2.8.3
scikit-learn==0.20.3
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
tensorflow-gpu==1.14.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.28.1
easydict==1.9
Cython==0.29.11
